Remove `BuiltinLintDiag::AbsPathWithModule`
